\name{rd_3_deprecated}
\alias{object.info}
\title{Deprecated}
\description{
This function is deprecated.

Please don't use it.

Use object.model() or object.summary() instead.
}
\usage{
object.info (object, n=6)
}
\arguments{
\item{object}{An arbitrary S3 object.}
\item{n}{Number of elements, lines or rows to print.}
}
\details{This function calls object.summary() with object and n.}
